requirements.txt
astroid==1.3.4
coverage==3.7.1
Django==1.8.5
django-nose==1.4.2
djangorestframework==3.3.0
logilab-common==1.1.0
nose==1.3.7
six==1.10.0
wheel==0.24.0
